Sprint #3 Deployment

+++ server/src/resolvers/equipmentInstanceResolver.ts
@@ -1,7 +1,8 @@
 import * as EquipmentRepo from "../repositories/Equipment/EquipmentRepository.js";
 import * as InstanceRepo from "../repositories/Equipment/EquipmentInstancesRepository.js";
+import * as RoomRepo from "../repositories/Rooms/RoomRepository.js"
 import { Privilege } from "../schemas/usersSchema.js";
-import { ApolloContext, ifAllowed } from "../context.js";
+import { ApolloContext } from "../context.js";
 import { EquipmentInstancesRow } from "../db/tables.js";
 import { createInstance, deleteInstance, getInstanceByID, getInstancesByEquipment, setInstanceName, setInstanceStatus } from "../repositories/Equipment/EquipmentInstancesRepository.js";
 import { createLog } from "../repositories/AuditLogs/AuditLogRepository.js";
@@ -37,24 +38,24 @@ const EquipmentInstanceResolver = {
     equipmentInstances: async (
       _parent: any,
       args: { equipmentID: number },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async () => {
+      { isStaff }: ApolloContext) =>
+      isStaff(async () => {
         return await getInstancesByEquipment(args.equipmentID)
       }),
 
     getInstanceByID: async (
       _parent: any,
       args: { id: number },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async () => {
+      { isStaff }: ApolloContext) =>
+      isStaff(async () => {
         return await getInstanceByID(args.id)
       }),
 
     getReaderPairedWithInstanceByInstanceId: async (
       _parent: any,
       args: { instanceID: number },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async () => {
+      { isStaff }: ApolloContext) =>
+      isStaff(async () => {
         return await InstanceRepo.getReaderByInstanceId(args.instanceID)
       }),
 
@@ -71,8 +72,8 @@ const EquipmentInstanceResolver = {
     createEquipmentinstance: async (
       _parent: any,
       args: { equipmentID: number, name: string },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async (user) => {
+      { isManager }: ApolloContext) =>
+      isManager(async (user) => {
         const equipment = await EquipmentRepo.getEquipmentByID(args.equipmentID);
         if (!equipment) throw new GraphQLError("Equipment does not exist");
         await createLog(`{user} created instance "${args.name}" on {equipment}`, "admin", { id: user.id, label: getUsersFullName(user) }, { id: equipment.id, label: equipment.name });
@@ -89,8 +90,8 @@ const EquipmentInstanceResolver = {
     updateInstance: async (
       _parent: any,
       args: { id: number, name: string, status: string, readerID: number | null },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async (user) => {
+      { isStaff }: ApolloContext) =>
+      isStaff(async (user) => {
 
         const instance = await InstanceRepo.getInstanceByID(args.id);
 
@@ -143,12 +144,16 @@ const EquipmentInstanceResolver = {
     setInstanceStatus: async (
       _parent: any,
       args: { id: number, status: string },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async (user) => {
+      { isStaff }: ApolloContext) =>
+      isStaff(async (user) => {
         const orig = await getInstanceByID(args.id);
         if (!orig) throw new GraphQLError("Instance does not exist");
         const equipment = await EquipmentRepo.getEquipmentByID(orig.equipmentID);
         if (!equipment) throw new GraphQLError("Equipment does not exist");
+        const room = await RoomRepo.getRoomByID(equipment.roomID);
+        if (!user.staff.includes(room?.zoneID ?? -1) && !user.manager.includes(room?.zoneID ?? -1) && !user.admin) {
+          throw new GraphQLError(`Not Privileged for Makerspace ${room?.zoneID}`);
+        }
         await createLog(`{user} changed instance "${orig.name}" status to "${args.status}" on {equipment}`, "admin", { id: user.id, label: getUsersFullName(user) }, { id: equipment.id, label: equipment.name });
         return await setInstanceStatus(args.id, args.status)
       }),
@@ -163,12 +168,16 @@ const EquipmentInstanceResolver = {
     setInstanceName: async (
       _parent: any,
       args: { id: number, name: string },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async (user) => {
+      { isManager }: ApolloContext) =>
+      isManager(async (user) => {
         const orig = await getInstanceByID(args.id);
         if (!orig) throw new GraphQLError("Instance does not exist");
         const equipment = await EquipmentRepo.getEquipmentByID(orig.equipmentID);
         if (!equipment) throw new GraphQLError("Equipment does not exist");
+        const room = await RoomRepo.getRoomByID(equipment.roomID);
+        if (!user.manager.includes(room?.zoneID ?? -1) && !user.admin) {
+          throw new GraphQLError(`Not Privileged for Makerspace ${room?.zoneID}`);
+        }
         await createLog(`{user} changed instance "${orig.name}" name to "${args.name}" on {equipment}`, "admin", { id: user.id, label: getUsersFullName(user) }, { id: equipment.id, label: equipment.name });
         return await setInstanceName(args.id, args.name)
       }),
@@ -182,20 +191,24 @@ const EquipmentInstanceResolver = {
     deleteInstance: async (
       _parent: any,
       args: { id: number },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async (user) => {
+      { isManager }: ApolloContext) =>
+      isManager(async (user) => {
         const orig = await getInstanceByID(args.id);
         if (!orig) throw new GraphQLError("Instance does not exist");
         const equipment = await EquipmentRepo.getEquipmentByID(orig.equipmentID);
         if (!equipment) throw new GraphQLError("Equipment does not exist");
+        const room = await RoomRepo.getRoomByID(equipment.roomID);
+        if (!user.manager.includes(room?.zoneID ?? -1) && !user.admin) {
+          throw new GraphQLError(`Not Privileged for Makerspace ${room?.zoneID}`);
+        }
         await createLog(`{user} deleted instance "${orig.name}" on {equipment}`, "admin", { id: user.id, label: getUsersFullName(user) }, { id: equipment.id, label: equipment.name });
         return await deleteInstance(args.id)
       }),
     assignReaderToEquipmentInstance: async (
       _parent: any,
       args: { instanceId: number, readerId: number | undefined },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async (user) => {
+      { isStaff }: ApolloContext) =>
+      isStaff(async (user) => {
 
         return InstanceRepo.assignReaderToEquipmentInstance(args.instanceId, args.readerId);
       }),
